PHPCS 3.x compat: add class aliases where possible
This file needs to be loaded quite early, but as PHPCS does not load the `bootstrap` file until *after* the ruleset has been parsed and the sniff classes autoloaded, we cannot use `bootstrap` to load the file.
So unfortunately, it will *have* to go into the any class file where the class extends an upstream parent class to prevent fatal errors from being thrown.